// apps/api/src/enums.ts
export enum Role {
  PASSENGER = 'PASSENGER',
  DRIVER = 'DRIVER'
}

export enum Area {
  BANANI = 'BANANI',
  GULSHAN_1 = 'GULSHAN_1',
  MOHAKHALI = 'MOHAKHALI',
  DHANMONDI = 'DHANMONDI',
  MIRPUR = 'MIRPUR',
  UTTARA = 'UTTARA',
  FARMGATE = 'FARMGATE',
  BASHUNDHARA = 'BASHUNDHARA'
}

export enum RideStatus {
  REQUESTED = 'REQUESTED',
  MATCHED = 'MATCHED',
  ACCEPTED = 'ACCEPTED',
  DRIVER_ARRIVED = 'DRIVER_ARRIVED',
  STARTED = 'STARTED',
  COMPLETED = 'COMPLETED',
  CANCELLED = 'CANCELLED'
}

export enum PoolStatus {
  OPEN = 'OPEN',
  ACCEPTED = 'ACCEPTED',
  DRIVER_ARRIVED = 'DRIVER_ARRIVED',
  STARTED = 'STARTED',
  COMPLETED = 'COMPLETED',
  CANCELLED = 'CANCELLED'
}

export enum PaymentMethod {
  CASH = 'CASH',
  TESLAPAY = 'TESLAPAY'
}